Merge 7.2 into feature/ai-appliance

+++ src/Raven.Server/Utils/AsyncBlittableJsonTextWriterForDebug.cs
@@ -179,7 +179,7 @@ namespace Raven.Server.Utils
             _inner.WritePropertyName(prop);
         }
 
-        public ValueTask<int> MaybeFlushAsync(CancellationToken token = default)
+        public ValueTask<long> MaybeFlushAsync(CancellationToken token = default)
         {
             return _inner.MaybeFlushAsync(token);
         }
